Refactor | #48 | @YongsHub | 애노테이션 패키지 분리
