+++ cakk-api/src/main/java/com/cakk/api/annotation/OperationDays.java
@@ -1,4 +1,4 @@
-package com.cakk.api.validator;
+package com.cakk.api.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,8 @@ import java.lang.annotation.Target;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
+
+import com.cakk.api.validator.OperationValidator;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
